Move getTSIGHashEnum to misc
